refactor(swap-service): extract pure helpers into swaps/utils.ts
Move standalone/pure helpers out of the service class:
- toSwap (relocated from types.ts — types.ts should only hold types)
- formatAmount
- getAffiliateCommissionRate
- estimateAffiliateFeeAmount
- resolveFeeAssetPrice

resolveSwapSellAmountUsd stays on the service (uses logger).
